backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
decorator==5.1.1
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
future==0.18.2
geocoder==1.38.1
idna==3.3
importlib-metadata==4.12.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2022.2.1
ratelim==0.1.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
djangorestframework-simplejwt==5.2.0